Add Display wrappers for FlakyTestRecord(s)
Mirrors the DisplayTestResultStats pattern: render the FLAKY M/T line via FlakyTestRecord::display(), and a slice via DisplayFlakyTestRecords::new(...). Removes the bespoke print_flaky_section in karva and the duplicated colored/format imports inside it; main now just write!s the wrapper.
